refactor individual service incl. css
