+++ src/app/domains/individual/feature-buttons/individual-buttons.widget.html
@@ -1,7 +1,7 @@
-<div class="individual-description-buttons">
-  <div class="individual-description-buttons__buttons">
+<div class="individual-buttons">
+  <div class="individual-buttons__buttons">
     <button
-      class="individual-description-buttons__button"
+      class="individual-buttons__button"
       mat-flat-button
       color="primary"
       type="button"
@@ -12,7 +12,7 @@
       {{ 'Objekt bearbeiten' | translate }}
     </button>
     <button
-      class="individual-description-buttons__button individual-description-buttons__button--secondary"
+      class="individual-buttons__button individual-buttons__button--secondary"
       mat-flat-button
       color="accent"
       type="button"
