+++ src/app/domains/individual/feature-description/individual-info.widget.html
@@ -1,26 +1,26 @@
-<div class="individual-description__basic-info">
+<div class="individual-info__basic-info">
   <div>
-    <div class="individual-description__label" data-test-id="individual_species">
+    <div class="individual-info__label" data-test-id="individual_species">
       {{ (species$ | async)?.de | translate }}
     </div>
-    <div class="individual-description__name" data-test-id="individual_name">{{ (individual$ | async).name }}</div>
+    <div class="individual-info__name" data-test-id="individual_name">{{ (individual$ | async).name }}</div>
   </div>
-  <div class="individual-description__label" *ngIf="(individual$ | async).last_observation_date">
+  <div class="individual-info__label" *ngIf="(individual$ | async).last_observation_date">
     <span translate>Letzte Beobachtung</span>
     <span>, </span>
     <span>{{ (individual$ | async).last_observation_date | shortdate }}</span>
   </div>
-  <div class="individual-description__name" *ngIf="lastPhenophase$ | async">
+  <div class="individual-info__name" *ngIf="lastPhenophase$ | async">
     <span
-      class="individual-description__phenophase-circle"
+      class="individual-info__phenophase-circle"
       [ngStyle]="{ 'background-color': lastPhenophaseColor$ | async }"
     ></span>
     <span>{{ (lastPhenophase$ | async)?.de | translate }}</span>
   </div>
   <div>
-    <span class="individual-description__label">{{ 'Von' | translate }} </span>
+    <span class="individual-info__label">{{ 'Von' | translate }} </span>
     <a
-      class="individual-description__creator-name"
+      class="individual-info__creator-name"
       [routerLink]="['/profile', (individual$ | async).user]"
       data-test-id="individual_creator"
       >{{ individualCreatorNickname$ | async }}
@@ -37,7 +37,7 @@
       <mat-icon svgIcon="ranger-badge"></mat-icon>
     </button>
   </div>
-  <div *ngIf="(individual$ | async)?.sensor as sensor" class="individual-description__sensor-badge">
+  <div *ngIf="(individual$ | async)?.sensor as sensor" class="individual-info__sensor-badge">
     <app-sensors-box [sensorLiveData]="sensor"></app-sensors-box>
   </div>
 </div>
