+++ src/app/domains/individual/feature-description/individual-info.widget.scss
@@ -16,7 +16,7 @@ button.navigation {
   width: 12.5rem;
 }
 
-.individual-description {
+.individual-info {
   &__label {
     @include detail-label;
   }
