+++ src/app/domains/individual/feature-description/masterdata-info.component.scss
@@ -4,7 +4,7 @@
 
 @import 'src/theming/mixins';
 
-.individual-description-basic-info {
+.masterdata-info {
   &__label {
     @include detail-label;
   }
