+++ src/app/domains/individual/geopos.service.ts
