+++ src/app/domains/individual/individual.page.html
@@ -5,16 +5,16 @@
       <app-individual-subscription-button [individual$]="detailSubject$"></app-individual-subscription-button>
     </div>
     <div class="individual-detail__content">
-      <app-individual-description-header
+      <app-individual-info
         *ngIf="detailSubject$ | async"
         class="individual-detail__description-header"
         [individual$]="detailSubject$"
-      ></app-individual-description-header>
-      <app-individual-description-basic-info
+      ></app-individual-info>
+      <app-masterdata-info
         *ngIf="detailSubject$ | async"
         class="individual-detail__basic-info"
         [individual$]="detailSubject$"
-      ></app-individual-description-basic-info>
+      ></app-masterdata-info>
       <div class="individual-detail__right_pane">
         <app-individual-observation-view
           class="individual-detail__observation-view"
@@ -22,14 +22,14 @@
           [isEditable$]="isEditable$"
           [isOwn$]="isOwn$"
         ></app-individual-observation-view>
-        <app-individual-description-buttons
+        <app-individual-buttons
           *ngIf="detailSubject$ | async"
           class="individual-detail__buttons"
           [individual$]="detailSubject$"
           [individualId]="detailId"
           [isEditable$]="isEditable$"
         >
-        </app-individual-description-buttons>
+        </app-individual-buttons>
       </div>
     </div>
   </div>
