+++ src/app/domains/individual/observation.service.ts
