--- src/app/domains/individual/station-species-phenophase-observations.model.ts
@@ -1,8 +1,0 @@
-import { Species } from '@shared/models/masterdata.model';
-import { PhenophaseObservation } from './station-phenophase-observation.model';
-
-export class SpeciesPhenophaseObservations {
-  treeId?: string;
-  species: Species;
-  phenophaseObservations: PhenophaseObservation[];
-}
